Auth midware moved from app to routes
