+++ services/secure-gateway/passThrough/passThrough_test.go
@@ -1,4 +1,4 @@
-package goThrough
+package passThrough
 
 /*
 import (
